Clean up code and add install target to Makefile

+++ tt.swift
@@ -6,21 +6,28 @@ if CommandLine.argc <= 1 {
 }
 
 let fm = FileManager.default
-let cwd = URL(fileURLWithPath: FileManager.default.currentDirectoryPath)
 var trashedPath: NSURL!
+var exitCode = 0
 
-for i: Int in 1..<CommandLine.arguments.count {
+for i in 1..<CommandLine.arguments.count {
     let argument = CommandLine.arguments[i]
     if argument.starts(with: "--") || argument.starts(with: "-") {
         print("Usage: tt <file>")
         exit(1)
     }
 
-    let filePath = URL(fileURLWithPath: argument, relativeTo: cwd)
+    let filePath = URL(fileURLWithPath: argument)
+    if !fm.fileExists(atPath: filePath.path()) {
+        print("tt: \(argument): No such file or directory")
+        exitCode = 1
+        continue
+    }
 
     do {
         try fm.trashItem(at: filePath, resultingItemURL: &trashedPath)
     } catch {
-        print("tt: \(argument): No such file or directory")
+        print("tt: \(error)")
     }
 }
+
+exit(Int32(exitCode))
